Move file functions into their own module.
